refactor: extract issue #1555 reject tests to separate file for CI line limit
The solve-queue.test.mjs file exceeded the 1500-line limit (1662 lines)
after adding issue #1555 tests. Extracted the 6 rejection tests to
tests/solve-queue-reject.test.mjs, bringing solve-queue.test.mjs to
1497 lines.
